refactor(rent): introduce rent service
