+++ src/main/java/com/medspace/application/usecase/payment/CreatePaymentUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.payment;
 
-import com.medspace.application.service.PaymentService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Payment;
 import com.medspace.domain.model.User;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -9,10 +9,9 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class CreatePaymentUseCase {
     @Inject
-    PaymentService paymentService;
+    RentService rentService;
 
     public Payment execute(Payment payment, User user) {
-        // TODO: Implement ownership/auth check after payment creation when RentAgreement is available
-        return paymentService.createPayment(payment);
+        return rentService.createPayment(payment);
     }
-} 
+}
