+++ src/main/java/com/medspace/application/usecase/payment/DeletePaymentByIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.payment;
 
-import com.medspace.application.service.PaymentService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.User;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -8,12 +8,12 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class DeletePaymentByIdUseCase {
     @Inject
-    PaymentService paymentService;
+    RentService rentService;
 
     public void execute(Long paymentId, User user) {
-        if (!paymentService.validatePaymentOwnership(paymentId, user.getId())) {
+        if (!rentService.validatePaymentOwnership(paymentId, user.getId())) {
             throw new SecurityException("User not authorized to delete this payment");
         }
-        paymentService.deletePaymentById(paymentId);
+        rentService.deletePaymentById(paymentId);
     }
-} 
+}
