+++ src/main/java/com/medspace/application/usecase/payment/GetPaymentByIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.payment;
 
-import com.medspace.application.service.PaymentService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Payment;
 import com.medspace.domain.model.User;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -9,12 +9,12 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class GetPaymentByIdUseCase {
     @Inject
-    PaymentService paymentService;
+    RentService rentService;
 
     public Payment execute(Long paymentId, User user) {
-        if (!paymentService.validatePaymentOwnership(paymentId, user.getId())) {
+        if (!rentService.validatePaymentOwnership(paymentId, user.getId())) {
             throw new SecurityException("User not authorized to view this payment");
         }
-        return paymentService.getPaymentById(paymentId);
+        return rentService.getPaymentById(paymentId);
     }
-} 
+}
