+++ src/main/java/com/medspace/application/usecase/payment/GetPaymentsByRentAgreementIdUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.payment;
 
-import com.medspace.application.service.PaymentService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Payment;
 import com.medspace.domain.model.User;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -10,13 +10,12 @@ import java.util.List;
 @ApplicationScoped
 public class GetPaymentsByRentAgreementIdUseCase {
     @Inject
-    PaymentService paymentService;
+    RentService rentService;
 
     public List<Payment> execute(Long rentAgreementId, User user) {
-        // TODO: Implement proper ownership check when RentAgreement is available
-        if (!paymentService.validatePaymentOwnership(null, user.getId())) {
+        if (!rentService.validatePaymentOwnership(null, user.getId())) {
             throw new SecurityException("User not authorized to view these payments");
         }
-        return paymentService.getPaymentsByRentAgreementId(rentAgreementId);
+        return rentService.getPaymentsByRentAgreementId(rentAgreementId);
     }
-} 
+}
