+++ src/main/java/com/medspace/application/usecase/payment/UpdatePaymentStatusUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.payment;
 
-import com.medspace.application.service.PaymentService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Payment;
 import com.medspace.domain.model.User;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -9,18 +9,18 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class UpdatePaymentStatusUseCase {
     @Inject
-    PaymentService paymentService;
+    RentService rentService;
 
     public Payment execute(Long paymentId, Payment.PaymentStatus newStatus, User user) {
         // Stubbed auth check
-        if (!paymentService.validatePaymentOwnership(paymentId, user.getId())) {
+        if (!rentService.validatePaymentOwnership(paymentId, user.getId())) {
             throw new SecurityException("User not authorized to update this payment");
         }
-        Payment payment = paymentService.getPaymentById(paymentId);
+        Payment payment = rentService.getPaymentById(paymentId);
         if (payment == null) {
             throw new IllegalArgumentException("Payment not found");
         }
         payment.setPaymentStatus(newStatus);
-        return paymentService.updatePayment(payment);
+        return rentService.updatePayment(payment);
     }
-} 
+}
