+++ src/main/java/com/medspace/application/usecase/rentRequest/CreateRentRequestDaysUseCase.java
@@ -1,7 +1,7 @@
 package com.medspace.application.usecase.rentRequest;
 
 import java.util.List;
-import com.medspace.application.service.RentRequestDayService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.RentRequestDay;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -10,12 +10,12 @@ import jakarta.transaction.Transactional;
 @ApplicationScoped
 public class CreateRentRequestDaysUseCase {
     @Inject
-    RentRequestDayService rentRequestDayService;
+    RentService rentService;
 
     @Transactional
     public void execute(List<RentRequestDay> rentRequestDays, Long rentRequestId) {
         for (RentRequestDay rentRequestDay : rentRequestDays) {
-            rentRequestDayService.createRentRequestDay(rentRequestDay, rentRequestId);
+            rentService.createRentRequestDay(rentRequestDay, rentRequestId);
         }
     }
 }
