+++ src/main/java/com/medspace/application/usecase/rentRequest/GetRentRequestsByLandlordIdUseCase.java
@@ -2,11 +2,11 @@ package com.medspace.application.usecase.rentRequest;
 
 import java.util.List;
 import com.medspace.application.service.ClinicService;
+import com.medspace.application.service.RentService;
 import com.medspace.application.service.UserService;
 import com.medspace.domain.model.Clinic;
 import com.medspace.domain.model.RentRequest;
 import com.medspace.domain.model.User;
-import com.medspace.domain.repository.RentRequestRepository;
 import com.medspace.infrastructure.dto.rentRequest.GetRentRequestPreviewDTO;
 import com.medspace.infrastructure.dto.rentRequest.RentRequestQueryFilterDTO;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -15,14 +15,14 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class GetRentRequestsByLandlordIdUseCase {
     @Inject
-    RentRequestRepository rentRequestRepository;
+    RentService rentService;
     @Inject
     UserService userService;
     @Inject
     ClinicService clinicService;
 
     public List<GetRentRequestPreviewDTO> execute(RentRequestQueryFilterDTO filterDTO) {
-        List<RentRequest> rentRequests = rentRequestRepository.findByLandlordId(filterDTO);
+        List<RentRequest> rentRequests = rentService.listRentRequestsByLandlordId(filterDTO);
 
         return rentRequests.stream().map(rentRequest -> {
             User tenant = userService.getUserById(rentRequest.getTenant().getId());
