+++ src/main/java/com/medspace/application/usecase/rentRequest/ListRentRequestDaysUseCase.java
@@ -1,7 +1,7 @@
 package com.medspace.application.usecase.rentRequest;
 
 import java.util.List;
-import com.medspace.application.service.RentRequestDayService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.RentRequestDay;
 import com.medspace.infrastructure.dto.rentRequest.GetRentRequestDayDTO;
 import jakarta.enterprise.context.ApplicationScoped;
@@ -10,11 +10,11 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class ListRentRequestDaysUseCase {
     @Inject
-    RentRequestDayService rentRequestDayService;
+    RentService rentService;
 
     public List<GetRentRequestDayDTO> execute(Long rentRequestId) {
         List<RentRequestDay> rentRequestDays =
-                rentRequestDayService.listRentRequestDaysByRentRequestId(rentRequestId);
+                rentService.listRentRequestDaysByRentRequestId(rentRequestId);
 
         return rentRequestDays.stream().map(GetRentRequestDayDTO::new).toList();
     }
