+++ src/main/java/com/medspace/application/usecase/rentRequest/ListRentRequestUseCase.java
@@ -1,18 +1,17 @@
 package com.medspace.application.usecase.rentRequest;
 
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.RentRequest;
-import com.medspace.domain.repository.RentRequestRepository;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import java.util.List;
 
 @ApplicationScoped
 public class ListRentRequestUseCase {
-
     @Inject
-    RentRequestRepository rentRequestRepository;
+    RentService rentService;
 
     public List<RentRequest> execute() {
-        return rentRequestRepository.findAllRequests();
+        return rentService.listAllRentRequests();
     }
 }
