+++ src/main/java/com/medspace/application/usecase/review/AssignReviewToAuthorUseCase.java
@@ -1,5 +1,6 @@
 package com.medspace.application.usecase.review;
-import com.medspace.application.service.ReviewService;
+
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Review;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -8,10 +9,10 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class AssignReviewToAuthorUseCase {
     @Inject
-    ReviewService reviewService;
+    RentService rentService;
 
     public Review execute(Long reviewId, Long authorId) {
-        return reviewService.assignAuthorToReview(reviewId, authorId);
+        return rentService.assignAuthorToReview(reviewId, authorId);
     }
 }
