+++ src/main/java/com/medspace/application/usecase/review/AssignReviewToClinicUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.review;
 
-import com.medspace.application.service.ReviewService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Review;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -9,14 +9,14 @@ import jakarta.transaction.Transactional;
 @ApplicationScoped
 public class AssignReviewToClinicUseCase {
     @Inject
-    ReviewService reviewService;
+    RentService rentService;
 
     @Transactional
     public Review execute(Long reviewId, Long clinicId) {
         if (clinicId == null) {
             throw new IllegalArgumentException("Clinic ID cannot be null");
         }
-        return reviewService.assignReviewToClinic(reviewId, clinicId);
+        return rentService.assignReviewToClinic(reviewId, clinicId);
     }
 
 }
