+++ src/main/java/com/medspace/application/usecase/review/CreateReviewUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.review;
 
-import com.medspace.application.service.ReviewService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Review;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -8,9 +8,9 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class CreateReviewUseCase {
     @Inject
-    ReviewService reviewService;
+    RentService rentService;
 
     public Review execute(Review review) {
-        return reviewService.createReview(review);
+        return rentService.createReview(review);
     }
 }
