+++ src/main/java/com/medspace/application/usecase/review/GetAllReviewsUseCase.java
@@ -1,6 +1,6 @@
 package com.medspace.application.usecase.review;
 
-import com.medspace.application.service.ReviewService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Review;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -10,9 +10,9 @@ import java.util.List;
 @ApplicationScoped
 public class GetAllReviewsUseCase {
     @Inject
-    ReviewService reviewService;
+    RentService rentService;
 
     public List<Review> execute() {
-        return reviewService.getAllReviews();
+        return rentService.getAllReviews();
     }
 }
