+++ src/main/java/com/medspace/application/usecase/review/GetReviewByIdUseCase.java
@@ -1,7 +1,7 @@
 package com.medspace.application.usecase.review;
 
 
-import com.medspace.application.service.ReviewService;
+import com.medspace.application.service.RentService;
 import com.medspace.domain.model.Review;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -9,9 +9,9 @@ import jakarta.inject.Inject;
 @ApplicationScoped
 public class GetReviewByIdUseCase {
     @Inject
-    ReviewService reviewService;
+    RentService rentService;
 
     public Review execute(Long id) {
-        return reviewService.getReviewById(id);
+        return rentService.getReviewById(id);
     }
 }
